fix : `help_func` import issue solved.
