+++ art/__init__.py
@@ -3,7 +3,7 @@
 from .errors import artError
 from .functions import aprint, art, randart
 from .functions import tprint, tsave, text2art
-from .functions import decor
+from .functions import decor, help_func
 from .functions import get_font_dic, set_default, art_list, font_list, decor_list
 from .params import ART_VERSION, FONT_NAMES, ART_NAMES, DECORATION_NAMES, DEFAULT_FONT
 from .params import ART_COUNTER, FONT_COUNTER, DECORATION_COUNTER
